requirements.txt
python-dotenv>=1.0.0
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
openai>=1.0.0
python-docx>=0.8.11
streamlit>=1.46.0 
google-generativeai
python-dotenv
pandas
tabulate
python-json-logger